+++ src/ast_printer/ast_printer.rs
@@ -1,11 +1,11 @@
 use crate::{
     ast_grammar::expr::{Expr, ExprVisitor},
-    scanner::token::{Literal, Token},
+    ast_grammar::token::{Literal, Token},
 };
 
 pub struct AstPrinter;
 impl AstPrinter {
-    pub fn print(expr: Expr) -> String {
+    pub fn _print(expr: Expr) -> String {
         return expr.accept(&Self);
     }
