--- src/interpreter/mod.rs
@@ -1,2 +1,0 @@
-pub mod interpreter;
-pub mod runtime_error;
