+++ src/main.rs
@@ -1,12 +1,10 @@
 use std::{env, process};
 
-use ast_grammar::expr::Expr;
-use scanner::token::{Literal, Token, TokenType};
-
-use crate::{ast_printer::ast_printer::AstPrinter, lox::Lox, rpn_printer::rpn_printer::RPNPrinter};
+use crate::lox::Lox;
 
 mod ast_grammar;
 mod ast_printer;
+mod error;
 mod interpreter;
 mod lox;
 mod parser;
