--- src/parser/mod.rs
@@ -1,2 +1,0 @@
-pub mod parse_error;
-pub mod parser;
