--- src/scanner/mod.rs
@@ -1,3 +1,0 @@
-pub mod error;
-pub mod scanner;
-pub mod token;
